backend/requirements.txt
# ======== Dependencias principales de Django ========
Django                      # Framework web principal para el backend
djangorestframework        # Biblioteca para crear APIs RESTful en Django
django-cors-headers        # Maneja los encabezados CORS para permitir solicitudes desde el frontend
djangorestframework-simplejwt  # Implementación de autenticación JWT para Django REST Framework

# ======== Soporte para desarrollo seguro ========
django-sslserver           # Servidor de desarrollo con soporte SSL para probar HTTPS localmente
django-extensions          # Extensiones útiles para Django (shell_plus, runserver_plus, etc.)
Werkzeug                   # Utilidades web usadas por django-extensions
pyOpenSSL                  # Biblioteca para trabajar con certificados SSL

# ======== Base de datos ========
psycopg2-binary            # Adaptador PostgreSQL para Django (versión precompilada)
psycopg                    # Versión 3 del adaptador PostgreSQL, usado por setup_db.py

# ======== Procesamiento de datos ========
gpxpy                      # Biblioteca para leer y analizar archivos GPX de rutas
pandas                     # Análisis de datos y manipulación para estadísticas
python-dateutil            # Funciones avanzadas para trabajar con fechas
pytz                       # Soporte para zonas horarias

# ======== Procesamiento de archivos XML ========
beautifulsoup4             # Analizador HTML/XML para procesar archivos TCX
lxml                       # Analizador XML rápido usado con BeautifulSoup

# ======== Seguridad ========
bcrypt                     # Biblioteca para cifrado seguro de contraseñas

# ======== Configuración y utilidades ========
python-dotenv              # Carga variables de entorno desde archivos .env
Pillow                     # Procesamiento de imágenes (para fotos de perfil)

# ======== Exportación de datos ========
reportlab                  # Generación de archivos PDF para informes de entrenamiento
python-pptx                # Creación de presentaciones PowerPoint
openpyxl                   # Manipulación de archivos Excel